requirements.txt
kubernetes==29.0.0
openai==1.30.5
PyYAML==6.0.1
rich==13.7.1
markdown2==2.4.13
requests==2.32.3